+++ src/components/Toast/Toast.tsx
@@ -1,65 +1,63 @@
-import NOOP from "lodash/noop";
-import React, { useMemo, useCallback, useRef } from "react";
-import PropTypes from "prop-types";
+import React, { FC, ReactElement, useCallback, useEffect, useMemo, useRef } from "react";
 import cx from "classnames";
 import { CSSTransition } from "react-transition-group";
 import Button from "../../components/Button/Button";
-import Icon from "../../components/Icon/Icon";
-import Check from "../../components/Icon/Icons/components/Check";
-import Alert from "../../components/Icon/Icons/components/Alert";
-import Info from "../../components/Icon/Icons/components/Info";
+import Icon, { IconSubComponentProps } from "../../components/Icon/Icon";
 import CloseSmall from "../Icon/Icons/components/CloseSmall";
 import ToastLink from "./ToastLink/ToastLink";
 import ToastButton from "./ToastButton/ToastButton";
-import { TOAST_TYPES, TOAST_ACTION_TYPES } from "./ToastConstants";
+import { ToastAction, ToastActionType, ToastType } from "./ToastConstants";
+import { getIcon } from "./ToastHelpers";
+import VibeComponentProps from "../../types/VibeComponentProps";
+import { NOOP } from "../../utils/function-utils";
 import "./Toast.scss";
 
-const defaultIconMap = {
-  [TOAST_TYPES.NORMAL]: Info,
-  [TOAST_TYPES.POSITIVE]: Check,
-  [TOAST_TYPES.NEGATIVE]: Alert
-};
+interface ToastProps extends VibeComponentProps {
+  actions?: ToastAction[];
+  /** If true, Toast is open (visible) */
+  open?: boolean;
+  type?: ToastType;
+  /** Possible to override the default icon */
+  icon?: string | React.FC<IconSubComponentProps> | null;
+  /** If true, won't show the icon */
+  hideIcon?: boolean;
+  /** The action to display */
+  action?: JSX.Element;
+  /** If false, won't show the close button */
+  closeable?: boolean;
+  onClose?: () => void;
+  /** The number of milliseconds to wait before
+   * automatically closing the Toast
+   * (0 or null cancels this behaviour) */
+  autoHideDuration?: number;
+  children?: ReactElement | ReactElement[] | string;
+}
 
-const getIcon = (type, icon) => {
-  /* icon may be node a may be a string */
-  if (icon && typeof icon === "object") {
-    return icon;
-  }
-  return icon || defaultIconMap[type] ? (
-    <Icon
-      iconType={icon ? Icon.type.ICON_FONT : Icon.type.SVG}
-      clickable={false}
-      icon={icon || defaultIconMap[type]}
-      iconSize="24px"
-      ignoreFocusStyle
-    />
-  ) : null;
-};
-
-const Toast = ({
-  open,
-  autoHideDuration,
-  type,
+const Toast: FC<ToastProps> & { types?: typeof ToastType; actionTypes?: typeof ToastActionType } = ({
+  open = false,
+  autoHideDuration = null,
+  type = ToastType.NORMAL,
   icon,
-  hideIcon,
+  hideIcon = false,
   action: deprecatedAction,
   actions,
   children,
-  closeable,
-  onClose,
+  closeable = true,
+  onClose = NOOP,
   className
 }) => {
   const toastLinks = useMemo(() => {
     return actions
       ? actions
-          .filter(action => action.type === TOAST_ACTION_TYPES.LINK)
+          .filter(action => action.type === ToastActionType.LINK)
           .map(({ type: _type, ...otherProps }) => <ToastLink key={otherProps.href} {...otherProps} />)
       : null;
   }, [actions]);
-  const toastButtons = useMemo(() => {
+
+  const toastButtons: JSX.Element[] | null = useMemo(() => {
     return actions
       ? actions
-          .filter(action => action.type === TOAST_ACTION_TYPES.BUTTON)
+          .filter(action => action.type === ToastActionType.BUTTON)
           .map(({ type: _type, content, ...otherProps }, index) => (
             <ToastButton key={`alert-button-${index}`} {...otherProps}>
               {content}
@@ -67,31 +65,35 @@ const Toast = ({
           ))
       : null;
   }, [actions]);
+
   const classNames = useMemo(
     () => cx("monday-style-toast", `monday-style-toast--type-${type}`, className),
     [type, className]
   );
+
   const handleClose = useCallback(() => {
     if (onClose) {
       onClose();
     }
   }, [onClose]);
+
   /* Timer */
-  const timerAutoHide = useRef();
+  const timerAutoHide = useRef<NodeJS.Timeout>();
   const setAutoHideTimer = useCallback(
-    duration => {
+    (duration: number) => {
       if (!onClose || duration == null) {
         return;
       }
 
       clearTimeout(timerAutoHide.current);
       timerAutoHide.current = setTimeout(() => {
-        handleClose(null, "timeout");
+        handleClose();
       }, duration);
     },
     [handleClose, onClose]
   );
-  React.useEffect(() => {
+
+  useEffect(() => {
     if (open && autoHideDuration > 0) {
       setAutoHideTimer(autoHideDuration);
     }
@@ -100,6 +102,7 @@ const Toast = ({
       clearTimeout(timerAutoHide.current);
     };
   }, [open, autoHideDuration, setAutoHideTimer]);
+
   const iconElement = !hideIcon && getIcon(type, icon);
 
   return (
@@ -134,41 +137,9 @@ const Toast = ({
   );
 };
 
-Toast.types = TOAST_TYPES;
-Toast.actionTypes = TOAST_ACTION_TYPES;
-
-Toast.propTypes = {
-  /** If true, Toast is open (visible) */
-  className: PropTypes.string,
-  actions: PropTypes.array,
-  open: PropTypes.bool,
-  type: PropTypes.oneOf([TOAST_TYPES.NORMAL, TOAST_TYPES.POSITIVE, TOAST_TYPES.NEGATIVE]),
-  /** Possible to override the dafult icon */
-  icon: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
-  /** If true, won't show the icon */
-  hideIcon: PropTypes.bool,
-  /** The action to display */
-  action: PropTypes.element,
-  /** If false, won't show the close button */
-  closeable: PropTypes.bool,
-  onClose: PropTypes.func,
-  /** The number of milliseconds to wait before
-   * automatically closing the Toast
-   * (0 or null cancels this behaviour) */
-  autoHideDuration: PropTypes.number
-};
-
-Toast.defaultProps = {
-  type: TOAST_TYPES.NORMAL,
-  className: "",
-  open: false,
-  action: null,
-  hideIcon: false,
-  icon: null,
-  closeable: true,
-  autoHideDuration: null,
-  actions: undefined,
-  onClose: NOOP
-};
+Object.assign(Toast, {
+  types: ToastType,
+  actionTypes: ToastActionType
+});
 
 export default Toast;